+++ Cargo.toml
@@ -7,5 +7,6 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-env_logger = "0.7.1"
-odbc = "0.17.0"
+sqlx = { version = "0.4.2", features = [ "mssql", "runtime-async-std-native-tls" ] }
+async-std = { version = "1.9.0", features = [ "attributes"] }
+futures = { version = "0.3.12" }
